WIP: Remove jersey deps from landscape service
