+++ .gitignore
@@ -5,14 +5,16 @@ build
 
 bin
 .classpath
+**/out/
 
 .project
+.idea/
 
 .DS_Store
 
 .gitmodules
 
-.settings/*
+**/.settings/*
 
 */.settings/*
 !*/.settings/edu.umd.cs.findbugs.core.prefs
