+++ build.gradle
@@ -4,6 +4,7 @@ plugins {
 
 allprojects {
   apply plugin: 'java-library'
+  apply plugin: "eclipse"
 }
 
 // defining project properties
@@ -106,16 +107,15 @@ subprojects {
     // Kafka
     implementation group: 'org.apache.kafka', name: 'kafka_2.12', version: '2.2.0'
     
-    if (project.name == 'user-service' || project.name == 'landscape-service' || project.name == 'discovery-service' || project.name == 'settings-service' || project.name == 'history-service' || project.name == 'broadcast-service') {
+    if (project.name == 'user-service' || project.name == 'discovery-service' || project.name == 'settings-service' || project.name == 'history-service' || project.name == 'broadcast-service') {
     	// subprojects, which are supposed to be a JAX-RS-based microservice
     
     	// Jersey and Servlet
 		implementation group: 'org.glassfish.jersey.containers', name: 'jersey-container-servlet', version:'2.27'
-		compile group: 'org.glassfish.jersey.core', name: 'jersey-client', version: '2.28'
+		implementation group: 'org.glassfish.jersey.core', name: 'jersey-client', version: '2.28'
 		implementation group: 'javax.servlet', name: 'javax.servlet-api', version:'3.0.1'
 		implementation group: 'org.glassfish.jersey.inject', name: 'jersey-hk2', version:'2.27'
-		// compile group: 'org.glassfish.jersey.test-framework.providers', name: 'jersey-test-framework-provider-jdk-http', version: '2.27'
-		testCompile group: 'org.glassfish.jersey.test-framework.providers', name: 'jersey-test-framework-provider-jetty', version: '2.27'
+		testImplementation group: 'org.glassfish.jersey.test-framework.providers', name: 'jersey-test-framework-provider-jetty', version: '2.27'
 		
 	  
 	    // JSON + Parsing
@@ -131,7 +131,7 @@ subprojects {
 		implementation group: 'org.glassfish.jersey.containers', name: 'jersey-container-jetty-servlet', version: '2.27'
 		
 		// Database
-		compile group: 'org.mongodb', name: 'mongo-java-driver', version: '3.9.1'
+		implementation group: 'org.mongodb', name: 'mongo-java-driver', version: '3.9.1'
 	}
