+++ landscape-service/src/main/java/net/explorviz/landscape/server/main/Main.java
@@ -1,11 +1,10 @@
 package net.explorviz.landscape.server.main;
 
-import net.explorviz.shared.config.helper.PropertyHelper;
-import org.eclipse.jetty.server.Server;
-import org.eclipse.jetty.servlet.ServletContextHandler;
-import org.eclipse.jetty.servlet.ServletHolder;
-import org.glassfish.jersey.server.ResourceConfig;
-import org.glassfish.jersey.servlet.ServletContainer;
+import net.explorviz.shared.common.provider.GenericTypeFinder;
+import net.explorviz.shared.landscape.model.helper.BaseEntity;
+import net.explorviz.shared.landscape.model.helper.TypeProvider;
+import org.glassfish.hk2.api.ServiceLocator;
+import org.glassfish.hk2.utilities.ServiceLocatorUtilities;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -15,7 +14,7 @@ public final class Main {
   private static final int DEFAULT_PORT = 8081;
 
   private Main() {
-    // Utility Class
+
   }
 
   /**
@@ -23,57 +22,16 @@ public final class Main {
    *
    */
   public static void main(final String[] args) {
-
-    final Server server = new Server(getPort());
-
-    final ServletHolder jerseyServlet = new ServletHolder(new ServletContainer(createJaxRsApp()));
-    final ServletContextHandler context = new ServletContextHandler(server, getContextPath());
-    context.addServlet(jerseyServlet, "/*");
-
-    try {
-      server.start();
-    } catch (final Exception e) { // NOPMD
-      LOGGER.error("Server start failed", e);
-    }
-
-    Runtime.getRuntime().addShutdownHook(new Thread(() -> {
-      try {
-        server.stop();
-      } catch (final Exception e) { // NOPMD
-        LOGGER.error("Server stop failed", e);
-      }
-    }));
+    // register Landscape Model classes, since we want to use them
+    // this must happen before initializing the ServiceLocator
+    TypeProvider.getExplorVizCoreTypesAsMap().forEach((classname, classRef) -> {
+      GenericTypeFinder.getTypeMap().put(classname, classRef);
+    });
+    ServiceLocator locator = ServiceLocatorUtilities.bind(new DependencyInjectionBinder());
+    LandscapeApplication app = locator.createAndInitialize(LandscapeApplication.class);
+    app.startExplorVizBackend();
   }
 
-  private static int getPort() {
-    try {
-      return PropertyHelper.getIntegerProperty("server.port");
-    } catch (final NumberFormatException e) {
-      if (LOGGER.isInfoEnabled()) {
-        LOGGER.info(
-            "ATTENTION: Using default port " + DEFAULT_PORT + ". Check explorviz.properties file.",
-            e);
-      }
-    }
-    return DEFAULT_PORT;
-  }
 
-  private static String getContextPath() {
-    final String statedContextPath = PropertyHelper.getStringProperty("server.contextPath");
-
-    if (statedContextPath == null) {
-      if (LOGGER.isInfoEnabled()) {
-        LOGGER.info("ATTENTION: Using default contextPath '/' for server. "
-            + "Maybe your stated server.contextPath property is no valid string.");
-      }
-      return "/";
-    } else {
-      return statedContextPath;
-    }
-  }
-
-  private static ResourceConfig createJaxRsApp() {
-    return new ResourceConfig(new LandscapeApplication());
-  }
 
 }
